Create Pitfalls, Surprises, and Shortcuts
PROJECT TITLE: "Grades" - 01.12 Pitfalls, Surprises, and Shortcuts

PURPOSE OF PROJECT: The purpose of the project is to find the average of 
the first 4 tests grades given. Then we must add 5 of our own tests grades and 
calculate the average for all those numbers. 

VERSION or DATE: August 11, 2020

HOW TO START THIS PROJECT: We are given local variables consisting of 3 ints and 
a double (for the average). We must use the correct arithmetic and increment 
operators in our calculations. In addition, there is a bit of casting needed in
this program in order to solve for the average (since some averages are decimal
values).

AUTHORS: Trisha Ganesh 

USER INSTRUCTIONS: The project will compute the average of the numbers given. 
It will also print the "Test Grade", "Total Points" and will move to the next test 
once it has calculated the average. For example, once finished with Test #1 it will
move to test #2. Hence the increment operator! 
-------------------------------PMR----------------------------------------
<+s>: This program allowed me to understand the use of the increment operator. 
In addition, I learned how to get the computer to calculate the average of given 
numbers. 

<-s>: I made a some mistakes with my print statements and I also had a few 
semicolon errors. At first, I had trouble with alignment and the use of "\t\t".
To fix this, I understood what my error was by using logic and reasoning. I said
my error out loud and this allowed me to fix my mistake.  
 
Overall, I had fun writing this program because I enjoyed calculating the 
average of a given set of numbers. With practice, I will be able to effectively 
write the program with confidence and without hesitation. 
--------------------------------------------------------------------------
In the future, I will be sure to check for mistakes such as the proper use of
"\t\t", "\n", println and print, and proper concatenation. 