Create Challenge Program
PROJECT TITLE: 02.10 Challenge Program

PURPOSE OF PROJECT: The purpose of this program is to print a receipt for the user based on the information they provided for the item they'd like to purchase. 

VERSION or DATE: 2/08/2021

HOW TO START THIS PROJECT: We are used to prompt the user (user input) to enter information. Next, we will conceal the information the user enters (such as credit/debit card info)

AUTHORS: Trisha Ganesh 

USER INSTRUCTIONS:

This program allows the potential user to see their receipt for their purchase. 
-------------------------------PMR----------------------------------------
<+s>: In this program, I understood how to use replace when replacing numbers with special characters. However, it was still unclear and I will need more practice. 

<-s>: With practice, I will be able to fully understand how to work this program. 
--------------------------------------------------------------------------
Understanding this info will help me in the future. 


I learned how to display the first letter of the user's first name using the material taught in class. I could improve a bit so I can do better in the future. 