Create Logical Operator
PROJECT TITLE: 03.07 Logical Operator

PURPOSE OF PROJECT: The purpose of this program is to calculate the TDEE given the BMI and an Activity Factor. 

VERSION or DATE: 2/07/2021

HOW TO START THIS PROJECT: We to display a table to the user and prompt them to enter their activity factor. We will then need to calculate the TDEE = BMR * AF from there. 

AUTHORS: Trisha Ganesh 

USER INSTRUCTIONS:

This project allows us to understand how TDEE is calculated given the BMR and AF. 
-------------------------------PMR----------------------------------------
<+s>: In this program, if, else if, else statements to calculate the AF to then calculate the TDEE. 

<-s>: Overall, I ran into some issues with Strings and converting; however, I asked Mr Shelton for helped and he explained it very well. 
--------------------------------------------------------------------------
I will be sure to remember this going forward. 

 

 

Updated: I prompted the user to enter their name, bmr, and gender. Next, you display the activity level diagram and you prompt the user to pick their choice of Activity Level. lastly, you print the results along with the calculated TDEE. (BMR*ActivityFactor)