Add files via upload
PROJECT TITLE: 05.05 Default Constructors

PURPOSE OF PROJECT: The purpose of this project is to implement an object for a new class. For example, GradeV3 was given as an example. I decided to use AgeV3. 

VERSION or DATE: 2/07/2021

HOW TO START THIS PROJECT: We have 2 methods above the main method. One that contains that formula to find the average of 2 numbers and the other to find the difference between the two. 

AUTHORS: Trisha Ganesh 

USER INSTRUCTIONS:

This project will allow us to the surface area formula given to us to calculate the average and difference between 2 numbers / ages. 

In this case, 11 and 15. 
-------------------------------PMR----------------------------------------
<+s>: This program allowed me to understand how to use multiple methods to solve a program. ( a mathematical problem to find the solution). 

<-s>: Overall, I enjoyed writing this program as it allowed me to work with multiple methods (aside from the main method). In addition, we needed to use default constructors.  
--------------------------------------------------------------------------
Understanding how default constructors and multiple methods were used in this program will help me going forward. 