Add files via upload
PROJECT TITLE: 05.08 Constructing Objects

PURPOSE OF PROJECT: The purpose of this project is to use private methods and combine it with other methods (including the main method) to calculate the average of 3 numbers.

VERSION or DATE: 2/07/2021

HOW TO START THIS PROJECT: We are supposed to have two files. One with the main method and the other with the calls and the private methods. This is to help us understand how private methods work. 

AUTHORS: Trisha Ganesh 

USER INSTRUCTIONS:

This project allows us to have three students with different ages (one from two years ago and the other is their current age). 
-------------------------------PMR----------------------------------------
<+s>: This program allowed me to understand how private methods work and how they can be used to work with other methods to solve a problem. 

<-s>: Overall, I ran into some issues such as getting to submit the other tester file for my program. However, it was a good practice for me. 
--------------------------------------------------------------------------
Understanding private methods will help me going forward. 